Separated validator and generator into separate packages.
